🤖 refactor(site/src/pages): nest remaining route pages
