+++ site/src/pages/GroupsPage/CreateGroupPage/CreateGroupPageView.stories.tsx
@@ -4,7 +4,7 @@ import { mockApiError } from "#/testHelpers/entities";
 import { CreateGroupPageView } from "./CreateGroupPageView";
 
 const meta: Meta<typeof CreateGroupPageView> = {
-	title: "pages/OrganizationGroupsPage/CreateGroupPageView",
+	title: "pages/GroupsPage/CreateGroupPage/CreateGroupPageView",
 	component: CreateGroupPageView,
 };
